requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
attrs==25.3.0
cffi==1.17.1
colorama==0.4.6
Columnar==1.4.1
frozenlist==1.7.0
idna==3.10
multidict==6.5.0
nextcord==2.6.0
propcache==0.3.2
pycparser==2.22
PyNaCl==1.5.0
python-dotenv==1.1.0
string-color==1.2.3
toolz==1.0.0
typing_extensions==4.14.0
wcwidth==0.2.13
yarl==1.20.1
